+++ app/StickersPack.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace App;
+
+use Illuminate\Database\Eloquent\Model;
+
+class StickersPack extends Model
+{
+    protected $table = 'stickers_packs';
+    public $primaryKey = 'name';
+    protected $keyType = 'string';
+
+    public function stickers()
+    {
+        return $this->hasMany('App\Sticker', 'pack', 'name');
+    }
+}