+++ app/Widgets/EmojisConfig/emojisconfig.js
@@ -1,0 +1,3 @@
+MovimWebsocket.attach(function () {
+    EmojisConfig_ajaxHttpGet();
+});